Remove ConsoleTestCase.php
This commit removes the ConsoleTestCase.php file which was no longer being used in the project. The file contained a class that extended from the TestCase class in the Orchestra\Testbench package and provided a method to run console commands. However, this functionality was not being utilized and the file was deemed unnecessary and removed.