Add prompt for missing migration name argument
This commit adds a prompt for missing migration name argument in AiMigrationCommand.php. If the name argument is not provided, the user will be prompted to enter a name. The entered name will then be transformed to snake case using Str::snake() method before being used. The prompt message is defined in the new promptForMissingArgumentsUsing() method.